+++ package.json
@@ -21,6 +21,7 @@
     "react-icons": "^4.2.0",
     "react-loading-overlay": "^1.0.1",
     "react-material-ui-form-validator": "^2.1.4",
+    "react-router": "^5.2.0",
     "react-scripts": "4.0.3",
     "react-scrollbar": "^0.5.6",
     "styled-components": "^5.2.3",
