+++ src/App.js
@@ -3,8 +3,8 @@ import React, {Component} from 'react';
 import SignIn from "./Component/User/SignIn";
 import SignUpConsultant from "./Component/User/SignUpConsultant";
 import SignUpUser from "./Component/User/SignUpUser";
+import EditProfile from './Component/User/Profile/EditProfile';
 import Profile from './Component/User/Profile/Profile';
-import Profile1 from './Component/User/Profile/Profile1';
 import MainPage from './Component/MainPage/MainPage';
 import Channels from './Component/MainPage/Channels';
 import Settings from './Component/MainPage/Settings';
@@ -26,8 +26,8 @@ class App extends Component {
                     <Route path="/CreateChannel" exact component={CreateChannel}/>
                     <Route path="/Channel/:channelId" children={<Dashboard title={'داشبورد'} />} component={Channel}/>
                     <Route path="/Dashboard" children={<Dashboard title={'داشبورد'} />} component={MainPage}/>
-                    <Route path="/Profile" children={<Dashboard title={'پروفایل'}/>} component={Profile}/>
-                    <Route path="/Profile1" children={<Dashboard title={'پروفایل'}/>} component={Profile1}/>
+                    <Route path="/EditProfile" children={<Dashboard title={'پروفایل'}/>} component={EditProfile}/>
+                    <Route path="/Profile/:username" children={<Dashboard title={'پروفایل'}/>} component={Profile}/>
                     <Route path="/Channels" children={<Dashboard title={'کانال ها'}/>} component={Channels}/>
                     <Route path="/GroupingChannel" children={<Dashboard title={'دسته بندی'}/>} component={GroupingChannel}/>
                      {/*<Route path="/Dashboard"  component={Dashboard}/>*/}
